+++ src/main/scala/org/podval/tools/node/Node.scala
@@ -1,6 +1,5 @@
 package org.podval.tools.node
 
-import org.podval.tools.platform.Exec
 import java.io.File
 
 final class Node(
@@ -13,18 +12,13 @@ final class Node(
   
   val nodeEnv: Seq[(String, String)] = Seq(("NODE_PATH", nodeModules.getAbsolutePath))
 
-  def node(arguments: String, log: String => Unit): String = run(
-    command = installation.node,
-    commandName = "node",
+  def node(arguments: String, log: String => Unit): String = installation.node(
     arguments,
-    cwd = None,
     extraEnv = nodeEnv,
     log
   )
 
-  def npm(arguments: String, log: String => Unit): String = run(
-    command = installation.npm,
-    commandName = "npm",
+  def npm(arguments: String, log: String => Unit): String = installation.npm(
     arguments,
     // in local mode, npm puts packages into node_modules under the current working directory
     cwd = Some(nodeModulesParent),
@@ -32,24 +26,6 @@ final class Node(
     extraEnv = nodeEnv ++ Seq(("PATH", installation.bin.getAbsolutePath + ":" + System.getenv("PATH"))),
     log
   )
-
-  private def run(
-    command: File,
-    commandName: String,
-    arguments: String,
-    cwd: Option[File],
-    extraEnv: Seq[(String, String)],
-    log: String => Unit
-  ): String =
-    log(s"Running '$commandName $arguments'...")
-    val output: String = Exec(
-      command,
-      arguments,
-      cwd,
-      extraEnv*
-    )
-    log(s"Output: [$output]\n")
-    output
 
   def setUpNodeProject(
     installModules: List[String],