+++ src/main/scala/org/podval/tools/scalajs/js/ClosureCompilerSourceMapper.scala
@@ -1,4 +1,4 @@
-package org.podval.tools.scalajs
+package org.podval.tools.scalajs.js
 
 import com.google.debugging.sourcemap.SourceMapConsumerV3
 import com.google.debugging.sourcemap.proto.Mapping.OriginalMapping